requirements.txt
# 基础依赖
pandas==2.1.4
numpy==1.26.2
requests==2.28.0
python-dotenv==1.0.0
ccxt==4.0.0
ta==0.10.0  # 技术分析库
loguru==0.7.2  # 日志管理
python-binance==1.0.17  # 币安API 
fastapi==0.105.0
uvicorn==0.24.0
websockets==12.0
aiohttp==3.9.1
python-multipart==0.0.6